Refactor tools Dockerfiles, docker_run.sh scripts (#1355)
