+++ README.md
@@ -223,13 +223,11 @@ root
 ├── <a href="./integration">integration</a>: Integration tests that spin up Obscuro networks.
 │   ├── <a href="./integration/simulation">simulation</a>: A series of tests that simulate running networks with different setups.
 ├── <a href="./testnet">testnet</a>: Utilities for deploying a testnet.
-└── <a href="./tools">tools</a>: Peripheral tooling. 
-    ├── <a href="./tools/azuredeployer">azuredeployer</a>: Automates deployment of Obscuro nodes to SGX-enabled Azure VMs.
-    ├── <a href="./tools/contractdeployer">contractdeployer</a>: Automates deployment of ERC20 and management contracts to the L1.
-    ├── <a href="./tools/networkmanager">networkmanager</a>: Network management tooling. Automates deployment of ERC20 and management contracts to the L1, and allows the injection of transactions into the network to simulate activity.
+└── <a href="./tools">tools</a>: Peripheral tooling.
+    ├── <a href="./tools/hardhatdeployer">hardhatdeployer</a>: Automates deployment of ERC20 and management contracts to the L1.
     ├── <a href="./tools/faucet">faucet</a>: Faucet for testnet.
     ├── <a href="./tools/obscuroscan">obscuroscan</a>: Tooling to monitor network transactions.
-    └── <a h ref="./tools/walletextension">walletextension</a>: Ensures sensitive messages to and from the Obscuro node are encrypted.
+    └── <a href="./tools/walletextension">walletextension</a>: Ensures sensitive messages to and from the Obscuro node are encrypted.
 
 </pre>
