+++ tools/hardhatdeployer/main/main.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"time"
 
+	contractdeployer "github.com/obscuronet/go-obscuro/tools/hardhatdeployer"
+
 	gethlog "github.com/ethereum/go-ethereum/log"
 	"github.com/obscuronet/go-obscuro/go/common/log"
-
-	"github.com/obscuronet/go-obscuro/tools/contractdeployer"
 )
 
 func main() {
