+++ tools/obscuroscan/Dockerfile
